requirements.txt
python-telegram-bot
qrcode
python-dotenv
